requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1637088766493/work
aiohttp @ file:///D:/bld/aiohttp_1637087174658/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1636093929600/work
argcomplete @ file:///tmp/build/80754af9/argcomplete_1618920853202/work
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor @ file:///home/conda/feedstock_root/build_artifacts/astor_1593610464257/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1610696312422/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
asynctest==0.13.0
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1640799537051/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
brotlipy @ file:///D:/bld/brotlipy_1636012439160/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1633010882559/work
certifi==2021.10.8
cffi @ file:///D:/bld/cffi_1636046306023/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
click @ file:///D:/bld/click_1645238363756/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cryptography @ file:///D:/bld/cryptography_1639699494546/work
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637091911212/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.3
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1617723079010/work
fonttools==4.29.1
frozenlist @ file:///D:/bld/frozenlist_1643222859261/work
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1596839682936/work
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1629296548061/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
google-pasta==0.2.0
grpcio @ file:///D:/bld/grpcio_1645230656486/work
h5py @ file:///D:/bld/h5py_1637964197172/work
haversine==2.5.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
imageio==2.16.1
imgaug==0.4.0
importlib-metadata @ file:///D:/bld/importlib-metadata_1646003465820/work
importlib-resources==5.4.0
ipykernel @ file:///C:/ci/ipykernel_1633527406033/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1643800131373/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.5
itsdangerous==2.1.1
jedi @ file:///C:/ci/jedi_1644297241925/work
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-console==6.4.3
jupyter-core @ file:///C:/ci/jupyter_core_1636537209059/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
kiwisolver==1.3.2
labelImg==1.8.6
lxml==4.8.0
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1637220118004/work
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune==0.8.4
multidict @ file:///D:/bld/multidict_1643055681390/work
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.6.3
notebook==6.4.8
numpy @ file:///D:/bld/numpy_1640083233759/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1643507977997/work
opencv-python==4.5.5.64
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1617859230218/work
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1638819640841/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyrsistent==0.18.1
PySocks @ file:///D:/bld/pysocks_1635862618120/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
PyWavelets==1.2.0
pywin32==302
pywinpty==2.0.5
pyzmq @ file:///C:/ci/pyzmq_1638435182681/work
qtconsole==5.2.2
QtPy==2.0.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1637781155505/work
scikit-image==0.19.2
scikit-learn==1.0.2
scipy @ file:///C:/bld/scipy_1637806996411/work
Send2Trash==1.8.0
Shapely==1.8.1.post1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.3.1
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1629725784641/work/tensorboard-2.5.0-py3-none-any.whl
tensorboard-data-server @ file:///D:/bld/tensorboard-data-server_1636045932656/work/tensorboard_data_server-0.6.0-py3-none-any.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.5.0
tensorflow-estimator @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-estimator_1623773901369/work/wheel_dir/tensorflow_estimator-2.5.0-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1641584929973/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1644332431572/work
widgetsnbextension==3.5.2
win-inet-pton @ file:///D:/bld/win_inet_pton_1635832912576/work
wincertstore==0.2
wrapt @ file:///D:/bld/wrapt_1635836573531/work
yarl @ file:///D:/bld/yarl_1636047058097/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1643828507773/work
